move exception definition
